feat(wai-handler-launch): Add function to launch arbitrary URLs
This commit adds the function runHostPortFullUrl to allow launching
arbitrary URLs instead of the hardcoded base http://127.0.0.1.
It allows using ::1 as the host IP address, while launching http://[::1]/,
where the additional brackets are required to access IPv6 IPs from the browser.

As the generation of the URL is moved into runHostPortUrl, both the C launch
function and the Haskell launch function can be reduced to (C)String -> IO ()
without breaking backwards compability.